Titweng_Backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
fastapi==0.118.0
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==11.3.0
polars==1.34.0
polars-runtime-32==1.34.0
propcache==0.4.0
psutil==7.1.0
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.2
sendgrid==6.12.5
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
twilio==9.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.205
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.37.0
Werkzeug==3.1.3
yarl==1.21.0
